pyproject.toml
######################################
###python-project-83/pyproject.toml###
######################################
[project]
name = "hexlet-code" 
version = "0.1.0"
dependencies = [
    "flask==3.0.0",
    "python-dotenv==1.0.0",
    "gunicorn==21.2.0"
]

[project.optional-dependencies]
dev = [
    "flake8==6.1.0",
    "black==23.12.0",
    "isort==5.13.2",
    "pytest==7.4.4", 
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
